Code - Web Application/cloud_functions/remote_sensing_data_fetcher/requirements.txt
pandas==2.2.1
google-cloud-storage==2.16.0
earthengine-api==0.1.397
google-cloud-pubsub==2.10.0
ordereddict